% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cell_growth_data}
\alias{cell_growth_data}
\title{Cancer cell growth data}
\format{A matrix with 27493 rows and 4 variables:
\describe{
\item{Timepoint}{Time / image number, from 0 to 8}
\item{X}{location coordinate X}
\item{Y}{location coordinate Y}
\item{Group}{group: 1, 2, 3}
}}
\usage{
cell_growth_data
}
\description{
A dataset containing location information of three types of cancer cells from 9 subsequent images
}
\keyword{datasets}
