% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copSTM.R
\name{copSTM}
\alias{copSTM}
\title{Fitting copula-based Multivariate spatio-temporal model}
\usage{
copSTM(x, y, K, n, maxit, eps = 0.1, std_err = F, B = 0,
  Message_prog = F)
}
\arguments{
\item{x}{A matrix of covariates generated by \code{make_data} or \code{sim_data}.}

\item{y}{A vector of count data response.}

\item{K}{The number of groups.}

\item{n}{An integer number, the image is tilled into n*n grids.}

\item{maxit}{Maximum number of iterations of maximum likelihood estimation.}

\item{eps}{Convergence threshold for maximum likelihood estimation.}

\item{std_err}{Logical, if TRUE returns bootstrapped standard errors of parameters.}

\item{B}{The number of bootstrap samples to be generated for computing the standard error of parameters. (Must be set if \code{std_err == TRUE})}

\item{Message_prog}{Logical, if TRUE, prints messages indicating progress. (Suggest on if long computational time is expected)}
}
\value{
A list with components
\itemize{
\item \code{likelihood:}   Maximized log-likelihood. 
\item \code{coefficients:} A list of parameter estimates: 
                    \itemize{
                      \item \code{intercepts:}  A separate intercept for each group,
                       \item \code{main_effects:}  Regression parameters indicating temporal impacts on growth between groups,
                       \item \code{correlations:} Spatial and multivariate correlations of observations at the same time point.}
\item \code{standard_error:}  A list of estimated standard errors, in the same format as \code{coefficients}. Returned only if \code{std_err == TRUE}. 
}
}
\description{
Fitting copula-based Multivariate spatio-temporal model
}
\examples{
### With simulated data
K <- 2
n <- 10
t_size <- 10
true_beta <- c(-1, 1, 0.5, -0.5, 0, 1)
true_rho <- c(-0.3, 0, 0.3, 0)
sim_dat <- sim_data(10, n, K, t_size, true_beta, true_rho) # in the same form returned by make_data
est_cop_sim <- copSTM(sim_dat$covariates, sim_dat$response, K, n, maxit = 50,
std_err = FALSE) # without standard error
print(est_cop_sim$coefficients)
\dontrun{
est_cop_sim <- copSTM(sim_dat$covariates, sim_dat$response, K, n, maxit = 50,
std_err = TRUE, B = 50, Message_prog = TRUE) # with standard error
print(est_cop_sim$standard_error)}

### with example data (expected around 40s)
\dontrun{
data("cell_growth_data") 
n <- 25
dat <- make_data(cell_growth_data, n)
K <- dat$K
est_cop_real <- copSTM(dat$covariates, dat$response, dat$K, n, maxit = 50)
print(est_cop_real$coefficients$main_effects)
cor0 <- matrix(NA, K, K)
cor1 <- matrix(NA, K, K)
diag(cor0) <- 1
diag(cor1) <- est_cop_real$coefficients$correlations[1:dat$K]
ind_r = K + 1
for(i in 1:(K - 1)){
  for(j in (i + 1):K){
    cor0[i, j] <- est_cop_real$coefficients$correlations[ind_r]
    ind_r = ind_r + 1
  }
}
for(i in 1:(K - 1)){
  for(j in (i + 1):K){
    cor1[i, j] <- est_cop_real$coefficients$correlations[ind_r]
    ind_r = ind_r + 1
  }
}
print("correlation in the same tile: ")
print(round(cor0, 3))
print("correlation in neighbouring tiles: ")
print(round(cor1, 3))
}

}
\seealso{
\code{\link{sim_data}}, \code{\link{make_data}}
}
