% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idpSTM.R
\name{idpSTM}
\alias{idpSTM}
\title{Fit spatio-temporal Model}
\usage{
idpSTM(data, n)
}
\arguments{
\item{data}{A matrix with four cloumns: time, x, y, group, where x and y are coordinates of location.}

\item{n}{An integer number indicating size of the grids. The image is tiled into an n*n grid. (Suggest n >= 6, otherwise neighbourhoods mostly overlapping with each other will lead to highly correlated covariates)}
}
\value{
A list with components
\itemize{
\item \code{likelihood:}   Maximized log-likelihood. 
\item \code{coefficients:} A list of parameter estimates: 
                    \itemize{
                      \item \code{intercepts:}  A separate intercept for each group,
                       \item \code{main_effects:}  Regression parameters indicating temporal impacts on growth between groups.}
\item \code{standard_error:}  A list of estimated standard errors, in the same format as \code{coefficients}. 
}
}
\description{
\code{idpSTM} takes a grouped data with information of spatial location and time point of each observation (typically a cell growth data), 
and fit a model of spatial neighbourhood impact between groups over time.
}
\examples{
n <- 25
data("cell_growth_data")
est <- idpSTM(cell_growth_data, n)
print("======= Estimates =======")
print(est$coefficients)
print("======= Standard errors =======")
print(est$standard_error)
}
