% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idpSTM.R
\name{idpSTModelSelect}
\alias{idpSTModelSelect}
\title{Model selsction of spatio-temporal model}
\usage{
idpSTModelSelect(data, n, ModelCnt, Message = F)
}
\arguments{
\item{data}{A matrix with four cloumns: time, x, y, group, where x and y are coordinates of location.}

\item{n}{An integer number indicating size of the grids. The image is tiled into an n*n grid. (Suggest n >= 6, otherwise neighbourhoods mostly overlapping with each other will lead to highly correlated covariates)}

\item{ModelCnt}{The number of models to be generated via Gibbs samplling.}

\item{Message}{Logical, if TRUE, prints the top 5 most frequently generated models.}
}
\value{
A list with components
\itemize{
\item \code{likelihood:}   Maximized log-likelihood. 
\item \code{coefficients:} A list of parameter estimates: 
                    \itemize{
                      \item \code{intercepts:}  A separate intercept for each group,
                       \item \code{main_effects:}  Regression parameters indicating temporal impacts on growth between groups,
                       }
\item \code{selected_model:}  A binary vector with 1 indicating selected variables and 0 otherwise.
\item \code{standard_error:}  A list of estimated standard errors, in the same format as \code{coefficients}. 
}
}
\description{
\code{idpSTModelSelect} performs variable selection on model \code{idpSTM} using BIC via Gibbs samplling.
}
\examples{
n <- 25
data("cell_growth_data") 
select_est <- idpSTModelSelect(cell_growth_data, 25, ModelCnt = 20, Message = FALSE)
print("======= Selected model =======")
print(select_est$selected_model)
print("======= Estimates =======")
print(select_est$coefficients)
print("======= Standard errors =======")
print(select_est$standard_error)
}
