% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idpSTM.R
\name{logGLMselect}
\alias{logGLMselect}
\title{Model selsction of log-GLM}
\usage{
logGLMselect(y, x, maxit = 50, skip = NULL, ModelCnt = 20,
  Message = T)
}
\arguments{
\item{y}{A vector of count data response.}

\item{x}{A matrix of covariates.}

\item{maxit}{Maximum number of iterations of maximum likelihood estimation.}

\item{skip}{A vector of indices of predictors to be forced in the selected model.}

\item{ModelCnt}{The number of models to be generated via Gibbs samplling.}

\item{Message}{Logical, if TRUE, prints the top 5 most frequently generated models. (Suggest on)}
}
\value{
A list with components
\itemize{
\item \code{likelihood:}   Maximized log-likelihood. 
\item \code{coefficients:} A vector of parameter estimates (the first one being intercept).
\item \code{selected_model:}  A binary vector with 1 indicating selected variables and 0 otherwise.
\item \code{standard_error:}  A vector of estimated standard errors. 
}
}
\description{
\code{logGLMselect} performs variable selection on log-GLM using BIC via Gibbs samplling.
}
\examples{
set.seed(777)  
p <- 10      # dimension
nn <- 3000   # sample size

x <- matrix(0, nn, p)
for(c in 1:p){
  x[, c] <- rnorm(nn, (p/5))
}
intercept <- 5
true_beta <- sample(c(-0.2, 0, 0.1), p, replace = TRUE)
y <- x \%*\% true_beta + intercept 
y <- sapply(y, exp)
y <- sapply(y, rpois, n = 1)

obj <- logGLMselect(y, x, skip = NULL, ModelCnt = 500, Message = TRUE)
cat("true model:", c(1, as.numeric(as.logical(true_beta))), "\\n")
}
