% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data.R
\name{make_data}
\alias{make_data}
\title{Data organization}
\usage{
make_data(data, n)
}
\arguments{
\item{data}{A matrix with four cloumns: time, x, y, group, where x and y are coordinates of location.}

\item{n}{An integer number indicating size of the grids. The image is tiled into an n*n grid. (Suggest n >= 6, otherwise neighbourhoods mostly overlapping with each other will lead to highly correlated covariates)}
}
\value{
A list with components: response, covariates, T (the number of time points) and K (the number of groups).
}
\description{
\code{make_data} takes a matrix of spatio-temporal data and returns a reorganized data ready for input of \code{copSTM} or \code{copSTModelSelect}.
}
\seealso{
\code{\link{copSTM_fit}}, \code{\link{copSTModelSelect}}
}
