% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copSTM.R
\name{copSTMSelect}
\alias{copSTMSelect}
\title{Model selection of copula-base multivariate spatio-temporal model}
\usage{
copSTMSelect(x, y, K, n, temporal, cor_type, marginal = "pois",
  ModelCnt = 100, B = 0, maxit1 = 50, maxit2 = 10,
  add_penalty = 0, Message_prog = TRUE, Message_res = TRUE,
  eps = 0.1)
}
\arguments{
\item{x}{A matrix of covariates, 
needs to be generated by \code{make_data} or \code{sim_data} if \code{temporal} == TRUE, 
otherwise could be any matrix.}

\item{y}{A vector of count data response. 
Required to be organized as 
\eqn{(y_{i, t, k}) = (y_{1, 1, 1}, \dots, y_{1, 1, K}, y_{2, 1, 1}, \dots, y_{2, 1, K}, 
\dots, y_{n^2, 1, K}, y_{1, 2, 1}, \dots, y_{n^2, T, K})}, 
where \eqn{(i, t, k)} denotes the indice of tile, time point and group respectively.}

\item{K}{The number of groups.}

\item{n}{An integer number, the image is tilled into n*n grids.}

\item{temporal}{Logical, if the data should be treated as temporal data. 
If TRUE, the output of regression coefficients takes the form of a K by K matrix, see \code{\link{idpSTM}} for details. 
Also the computation of standard errors requires bootstrapping, see argument \code{std_err} for details.}

\item{cor_type}{Correlation type. 
\itemize{
\item \code{sp}: Spatial only, consider only correlation of the same group from neighbouring tiles,
\item \code{mv}: Multivariate only, consider only correlation between different groups in the same tile,
\item \code{both}: Both "sp" and "mv", as well as between-group correlation from neighbouring tiles,
\item \code{ind}: Independent, \code{copSTM} is equivalent to GLM, please enter K = 1 and n = 1.
}}

\item{marginal}{Marginal distribution: "pois" for Poisson, "nbinom" for negative binomial.}

\item{ModelCnt}{The number of models to be generated via Gibbs samplling (default 100).}

\item{B}{The number of bootstrap samples to be generated for computing the penalty in CL-BIC and standard error of parameters.}

\item{maxit1}{The maximum number of iterations of maximum composite likelihood estimation for full model (default 50).}

\item{maxit2}{The maximum number of iterations of maximum composite likelihood estimation for sub-models,
expect to be much smaller than maxit1, since submodels take estimates from the full model as initial values (default 10).}

\item{add_penalty}{A tunning parameter to enforce sparsity. The larger this value is, the more favorable a sparse model becomes.}

\item{Message_prog}{Logical, if TRUE, prints messages indicating progress. (Suggest on if long computational time is expected)}

\item{Message_res}{Logical, if TRUE, prints the top 5 most frequently generated models.}

\item{eps}{Convergence threshold for maximum likelihood estimation (default 0.1).}
}
\value{
A list with components
\itemize{
\item \code{likelihood:}   Maximized log-likelihood. 
\item \code{coefficients:} A list of parameter estimates: 
                    \itemize{
                      \item \code{intercepts:}  A separate intercept for each group,
                       \item \code{main_effects:}  Regression coefficients, 
                       \item \code{correlations:} Spatial and multivariate correlations of observations at the same time point, 
                       \item \code{dispersion:} Overdispersion parameter, returned only when \code{marginal == "nbinom"}.}
\item \code{selected_model:}  A binary vector with 1 indicating selected variables and 0 otherwise.
\item \code{standard_error:}  A list of estimated standard errors, in the same format as \code{coefficient}. Returned only if \code{std_err == TRUE}. 
}
}
\description{
\code{copSTMSelect} finds the best model among all possible models. 
Models are fitted with \code{copSTM}, 
marginal distribution can be specified as Poisson or Negative binomial. 
Models are ranked with the composite likelihood BIC (CL-BIC) proposed by Gao and Song (2010). 
The best models are found using a Gibbs samplling method
introduced by Qian and Field (2002), which allows very large candidate sets to be adressed.
}
\examples{
\dontrun{
## Ex1.
set.seed(444)
K <- 2  # number of groups
n <- 10 # (squared root of) number of tiles
t_size <- 10 # number of time points
true_beta <- c(-1, 1, 0.5, -0.5, 0, 1)
true_rho <- c(-0.3, 0, 0.3, 0)
sim_dat <- sim_data(n, K, t_size, true_beta, marginal = "pois", 
                     temporal = TRUE, rho = true_rho, cor_type = "both")
# generate data in the same form returned by make_data

MdCnt <- 50
B <- 20 
gm <- 0.5  
set.seed(444)
start <- Sys.time()  # expected less than 2 mins 
select_est <- copSTMSelect(sim_dat$covariates, sim_dat$response, K = K, n = n, 
                            marginal = "pois", temporal = FALSE, cor_type = "both", 
                            ModelCnt = MdCnt, B = B, add_penalty = gm,
                            Message_prog = TRUE, Message_res = TRUE, eps = 0.1)
finish <- Sys.time()
print(finish - start)
as.numeric(as.logical(c(true_beta, true_rho)))
}

\dontrun{
### Ex2. (expect 3 mins)
start <- Sys.time()
set.seed(444)
K <- 2  # number of groups
n <- 10 # (squared root of) number of tiles
t_size <- 50 # number of time points
true_beta <- c(1, 0.8, -0.6, 0.4, 0, -0.2)
true_rho <- c(-0.3, 0, 0.3, 0)
dispersion <- 1
sim_dat <- sim_data(n, K, temporal = FALSE, t_size, marginal = "nbinom",
                    dispersion = dispersion, true_beta, true_rho, cor_type = "both") 
                    
MdCnt <- 20
gm <- 0.5 
start <- Sys.time()  # expected less than 2 mins 
select_est <- copSTMSelect(sim_dat$covariates, sim_dat$response, K = K, n = n, 
                            marginal = "nbinom", temporal = FALSE, cor_type = "both", 
                            ModelCnt = MdCnt, add_penalty = gm,
                            Message_prog = TRUE, Message_res = TRUE)
finish <- Sys.time()
print(finish - start)
as.numeric(as.logical(c(true_beta, true_rho, 1)))
print(select_est$coefficients)
}

}
\references{
{
X. Gao and P. X.-K. Song. Composite likelihood bayesian information criteria for model selectionin 
high-dimensional data.Journal of the American Statistical Association, 105(492):1531–1540,2010.
}

{
G. Qian and C. Field. Using mcmc for logistic regression model selection involving large
number of candidate models. In Monte Carlo and Quasi-Monte Carlo Methods 2000, pages 460–474. Springer, 2002.
}
}
