% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data.R
\name{sim_data}
\alias{sim_data}
\title{Simulate spatio-temporal data}
\usage{
sim_data(y_ini, n, K, t_size, beta, rho_v)
}
\arguments{
\item{y_ini}{An integer initial value for the first time point.}

\item{n}{An integer spatial parameter. Data is simulated as n*n grids.}

\item{K}{The number of groups.}

\item{t_size}{The number time points to be generated.}

\item{beta}{True values of regression parameters.}

\item{rho_v}{True values of correlations.}
}
\value{
Data ready to input to \code{copSTM} or \code{copSTModelSelect}.
}
\description{
Simulate data for testing \code{copSTM} and \code{copSTModelSelect}.
}
\seealso{
\code{\link{copSTM_fit}}, \code{\link{copSTModelSelect}}
}
